+++ changelog.md
@@ -1,4 +1,8 @@
-### 2022-11-14 一23:39:15  
+### 2022-11-15 二 10:30:17 
+- add monokai theme
+
+
+### 2022-11-14 一 23:39:15  
 - add lua/nvim_tokyonight.lua plugin (theme plugin)
 - lspsaga.nvim NeoVim Builtin LSP 增强插件 http://xfyuan.github.io/2021/03/a-awesome-neovim-lsp-plugin/
 - add alpha-nvim plugin ()